refactor unit test for TravisCI
