+++ .travis.yml
@@ -10,7 +10,7 @@ addons:
       - libtokyocabinet-dev
 
 rvm:
-  - 2.1.0
+  - 2.1.6
   - 2.2.3
   - ruby-head
