+++ test/t_mapcountplugin.rb
@@ -8,6 +8,7 @@ Roma::Client::RomaClient.class_eval{
 }
 
 class MapCountPluginTest < Test::Unit::TestCase
+  self.test_order = :defined
   include RomaTestUtils
 
   DATE_FORMAT = "%Y-%m-%dT%H:%M:%S +00"
@@ -25,6 +26,16 @@ class MapCountPluginTest < Test::Unit::TestCase
     stop_roma
   end
 
+  def check_roma_value(h, res)
+    h.each{|k, v|
+      if k == 'last_updated_date'
+        assert_equal true, res[k].between?(v, v+1)
+      else
+        assert_equal v, res[k]
+      end
+    }
+  end
+
   def test_mapcount_get(n = 5)
     n.times do |i|
       @sk = "subkey#{i}"
@@ -34,16 +45,16 @@ class MapCountPluginTest < Test::Unit::TestCase
       (i+1).times do |j|
         retc["subkey#{j}"] = 1
       end
-      assert_equal retc, @rc.mapcount_countup(@k, @sk, 0)
+      check_roma_value(retc, @rc.mapcount_countup(@k, @sk, 0))
       
       retg = {"last_updated_date"=>@lt}
       retg["subkey#{i}"] = 1
-      assert_equal retg, @rc.mapcount_get(@k, @sk)
-      assert_equal retc, @rc.mapcount_get(@k)
+      check_roma_value(retg, @rc.mapcount_get(@k, @sk))
+      check_roma_value(retc, @rc.mapcount_get(@k))
     end
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_get(@k, 'subkey0,subkey1')
+    check_roma_value(ret2, @rc.mapcount_get(@k, 'subkey0,subkey1'))
   end
 
   def test_mapcount_countup_array
@@ -52,30 +63,30 @@ class MapCountPluginTest < Test::Unit::TestCase
 
   def test_mapcount_countup_array_multi
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_countup(@k, ["subkey0", "subkey1"])
+    check_roma_value(ret2, @rc.mapcount_countup(@k, ["subkey0", "subkey1"]))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>2}
     @rc.mapcount_countup(@k, ["subkey1"])
-    assert_equal ret2, @rc.mapcount_get(@k)
+    check_roma_value(ret2, @rc.mapcount_get(@k))
   end
 
   def test_mapcount_countup_hash
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_countup(@k, ["subkey0", "subkey1"])
+    check_roma_value(ret2, @rc.mapcount_countup(@k, ["subkey0", "subkey1"]))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>4, "subkey1"=>1}
     @rc.mapcount_countup(@k, {"subkey0" => 3})
-    assert_equal ret2, @rc.mapcount_get(@k)
+    check_roma_value(ret2, @rc.mapcount_get(@k))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>4, "subkey1"=>11}
     @rc.mapcount_countup(@k, {"subkey1" => 10})
-    assert_equal ret2, @rc.mapcount_get(@k)
+    check_roma_value(ret2, @rc.mapcount_get(@k))
   end
 
   def test_mapcount_countup_degits
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>125, "subkey1"=>30}
     @rc.mapcount_countup(@k, {"subkey0" => 125, "subkey1" => 30})
-    assert_equal ret2, @rc.mapcount_get(@k)
+    check_roma_value(ret2, @rc.mapcount_get(@k))
   end
 
   def test_mapcount_countup_expt
@@ -90,29 +101,29 @@ class MapCountPluginTest < Test::Unit::TestCase
     ret_time = {"last_updated_date"=>@lt}
 
     assert_nil @rc.mapcount_update(@k)
-    assert_equal @ret, @rc.mapcount_countup(@k, @sk, 0)
-    assert_equal @ret, @rc.mapcount_update(@k)
-    assert_equal ret_time, @rc.mapcount_update(@k, 'subkey2')
-    assert_equal @ret, @rc.mapcount_update(@k, nil, 1)
+    check_roma_value(@ret, @rc.mapcount_countup(@k, @sk, 0))
+    check_roma_value(@ret, @rc.mapcount_update(@k))
+    check_roma_value(ret_time, @rc.mapcount_update(@k, 'subkey2'))
+    check_roma_value(@ret, @rc.mapcount_update(@k, nil, 1))
     sleep 2
     assert_nil @rc.mapcount_get(@k)
   end
 
   def test_mapcount_update_subkeys
     ret = {"last_updated_date"=>@lt, @sk=>1}
-    assert_equal @ret, @rc.mapcount_countup(@k, @sk, 0)
-    assert_equal @ret, @rc.mapcount_update(@k, @sk, 1)
+    check_roma_value(@ret, @rc.mapcount_countup(@k, @sk, 0))
+    check_roma_value(@ret, @rc.mapcount_update(@k, @sk, 1))
     sleep 2
     assert_nil @rc.mapcount_get(@k)
   end
 
   def test_mapcount_update_time
     ret = {"last_updated_date"=>@lt, @sk=>1}
-    assert_equal @ret, @rc.mapcount_countup(@k, @sk, 0)
+    check_roma_value(@ret, @rc.mapcount_countup(@k, @sk, 0))
     sleep 2
     lt2 = Time.parse(Time.now.strftime(DATE_FORMAT))
     ret2 = {"last_updated_date"=>lt2, @sk=>1}
-    assert_equal ret2, @rc.mapcount_update(@k)
+    check_roma_value(ret2, @rc.mapcount_update(@k))
   end
 
   def test_counts
@@ -156,16 +167,16 @@ class MapCountPluginTest < Test::Unit::TestCase
       (i+1).times do |j|
         retc["subkey#{j}"] = 1
       end
-      assert_equal retc, @rc.mapcount_countup_ms(@k, @sk, 0)
-      
+      check_roma_value(retc, @rc.mapcount_countup_ms(@k, @sk, 0))
+
       retg = {"last_updated_date"=>@lt}
       retg["subkey#{i}"] = 1
-      assert_equal retg, @rc.mapcount_get_ms(@k, @sk)
-      assert_equal retc, @rc.mapcount_get_ms(@k)
+      check_roma_value(retg, @rc.mapcount_get_ms(@k, @sk))
+      check_roma_value(retc, @rc.mapcount_get_ms(@k))
     end
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_get_ms(@k, 'subkey0,subkey1')
+    check_roma_value(ret2, @rc.mapcount_get_ms(@k, 'subkey0,subkey1'))
   end
 
   def test_mapcount_countup_ms_array
@@ -174,30 +185,30 @@ class MapCountPluginTest < Test::Unit::TestCase
 
   def test_mapcount_countup_ms_array_multi
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_countup_ms(@k, ["subkey0", "subkey1"])
+    check_roma_value(ret2, @rc.mapcount_countup_ms(@k, ["subkey0", "subkey1"]))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>2}
     @rc.mapcount_countup_ms(@k, ["subkey1"])
-    assert_equal ret2, @rc.mapcount_get_ms(@k)
+    check_roma_value(ret2, @rc.mapcount_get_ms(@k))
   end
 
   def test_mapcount_countup_ms_hash
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>1, "subkey1"=>1}
-    assert_equal ret2, @rc.mapcount_countup_ms(@k, ["subkey0", "subkey1"])
+    check_roma_value(ret2, @rc.mapcount_countup_ms(@k, ["subkey0", "subkey1"]))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>4, "subkey1"=>1}
     @rc.mapcount_countup_ms(@k, {"subkey0" => 3})
-    assert_equal ret2, @rc.mapcount_get_ms(@k)
+    check_roma_value(ret2, @rc.mapcount_get_ms(@k))
 
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>4, "subkey1"=>11}
     @rc.mapcount_countup_ms(@k, {"subkey1" => 10})
-    assert_equal ret2, @rc.mapcount_get_ms(@k)
+    check_roma_value(ret2, @rc.mapcount_get_ms(@k))
   end
 
   def test_mapcount_countup_ms_degits
     ret2 = {"last_updated_date"=>@lt, "subkey0"=>125, "subkey1"=>30}
     @rc.mapcount_countup_ms(@k, {"subkey0" => 125, "subkey1" => 30})
-    assert_equal ret2, @rc.mapcount_get_ms(@k)
+    check_roma_value(ret2, @rc.mapcount_get_ms(@k))
   end
 
   def test_mapcount_countup_ms_expt
@@ -212,29 +223,29 @@ class MapCountPluginTest < Test::Unit::TestCase
     ret_time = {"last_updated_date"=>@lt}
 
     assert_nil @rc.mapcount_update_ms(@k)
-    assert_equal @ret, @rc.mapcount_countup_ms(@k, @sk, 0)
-    assert_equal @ret, @rc.mapcount_update_ms(@k)
-    assert_equal ret_time, @rc.mapcount_update_ms(@k, 'subkey2')
-    assert_equal @ret, @rc.mapcount_update_ms(@k, nil, 1)
+    check_roma_value(@ret, @rc.mapcount_countup_ms(@k, @sk, 0))
+    check_roma_value(@ret, @rc.mapcount_update_ms(@k))
+    check_roma_value(ret_time, @rc.mapcount_update_ms(@k, 'subkey2'))
+    check_roma_value(@ret, @rc.mapcount_update_ms(@k, nil, 1))
     sleep 2
     assert_nil @rc.mapcount_get_ms(@k)
   end
 
   def test_mapcount_update_ms_subkeys
     ret = {"last_updated_date"=>@lt, @sk=>1}
-    assert_equal @ret, @rc.mapcount_countup_ms(@k, @sk, 0)
-    assert_equal @ret, @rc.mapcount_update_ms(@k, @sk, 1)
+    check_roma_value(@ret, @rc.mapcount_countup_ms(@k, @sk, 0))
+    check_roma_value(@ret, @rc.mapcount_update_ms(@k, @sk, 1))
     sleep 2
     assert_nil @rc.mapcount_get_ms(@k)
   end
 
   def test_mapcount_update_ms_time
     ret = {"last_updated_date"=>@lt, @sk=>1}
-    assert_equal @ret, @rc.mapcount_countup_ms(@k, @sk, 0)
+    check_roma_value(@ret, @rc.mapcount_countup_ms(@k, @sk, 0))
     sleep 2
     lt2 = Time.parse(Time.now.strftime(DATE_FORMAT))
     ret2 = {"last_updated_date"=>lt2, @sk=>1}
-    assert_equal ret2, @rc.mapcount_update_ms(@k)
+    check_roma_value(ret2, @rc.mapcount_update_ms(@k))
   end
 
   def test_counts_ms
